Simplified EC patch
The existing patch replaced the entire content of the upstream ChromeOS
EC getversion.sh with a single line.  This is brittle, as any change to
the upstream file will cause failed of the patch to apply going forware.

This change modifies the patch to insert two lines at the very top of
the script, which does our logic and exits without the bulk of the
script ever running.  That way, most changes in the upstream file will
not interfere with the ability of the patch to continue to apply.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>