Update EC commit, support for SPI1, blue button
Add SPI1 as another SPI port, in addition to SPI2 and QSPI.  Unlike
existing SPI and I2C ports, the pins of SPI1 default to being GPIO, and
must explicitly be configured as "Alternate" mode, in order for SPI to
be used on them.  This is done to maintain compatibility with existing
uses of the pins.

Pressing the blue button on HyperDebug will pull a configurable "reset"
pin low (provided it is configured to have output enabled), default is
CN10_29.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>