Update EC commit, binary GPIO monitoring protocol
HyperDebug has memory to buffer tens of thousands of GPIO edge events,
when using the `GpioMonitoring` functionality.  However, transferring
them to opentitantool via the text console protocol over USB was
inefficient.  This change includes a new "vendor extension" to the
binary CMSIS-DAP protocol, to be used for efficiently retrieving many
events (basically a memory dump of the internal buffer.)

Signed-off-by: Jes B. Klinke <jbk@chromium.org>