Update EC commit, advanced bit-banging
Add support for the `await()` feature in bitbanging pattern, allowing
waveform to stop a particular points, waiting for certain inputs before
proceeding.  This is useful e.g. for bit-banging an I2C host, when the
I2C device may stretch the clock, or for generating waveforms are
precise timing relative to some trigger from the chip under test.

34f0cfdfcc board/hyperdebug: Bit bang support for triggers
6e10acec2f board/hyperdebug: No-op refactoring
0c6e97ee41 board/hyperdebug: Central declaration of timer assignment
6e9838a098 board/hyperdebug: Declare pins on CN11 and CN12