Adapt to Makefile changes in ChromeOS EC repository
The build system has been slightly changed, requiring a slight change in
how to specify the path to the cross compiler.

Also, the EC codebase now assumes GCC 11 or above, while even the newest
lowRISC/crt provides gcc 10.3.1.  So a small patch was required, as the
older GCC does not recognize the `min-pagesize` parameter for
controlling GCC warnings.